Fix furnace BurnTime overflow (Blast Furnace & Smoker)

+++ src/main/java/ganymedes01/etfuturum/tileentities/TileEntityBlastFurnace.java
@@ -115,7 +115,7 @@ public class TileEntityBlastFurnace extends TileEntityFurnace implements ISidedI
 		this.furnaceItemStacks = new ItemStack[this.getSizeInventory()];
 		Utils.loadItemStacksFromNBT(nbttaglist, this.furnaceItemStacks);
 
-		this.furnaceBurnTime = compound.getShort("BurnTime");
+		this.furnaceBurnTime = compound.getInteger("BurnTime");
 		this.furnaceCookTime = compound.getShort("CookTime");
 		this.currentItemBurnTime = getItemBurnTime(this.furnaceItemStacks[1]);
 
@@ -127,7 +127,7 @@ public class TileEntityBlastFurnace extends TileEntityFurnace implements ISidedI
 	@Override
 	public void writeToNBT(NBTTagCompound compound) {
 		super.writeToNBT(compound);
-		compound.setShort("BurnTime", (short) this.furnaceBurnTime);
+		compound.setInteger("BurnTime", this.furnaceBurnTime);
 		compound.setShort("CookTime", (short) this.furnaceCookTime);
 
 		compound.setTag("Items", Utils.writeItemStacksToNBT(this.furnaceItemStacks));
